Keep only unit tests in unittests
