+++ tests/unittests/scenario_builder/systems/with_scenarization.yml
@@ -1,0 +1,44 @@
+# Copyright (c) 2024, RTE (https://www.rte-france.com)
+#
+# See AUTHORS.txt
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+# SPDX-License-Identifier: MPL-2.0
+#
+# This file is part of the Antares project.
+system:
+  model-libraries: basic
+  nodes:
+    - id: N
+      model: basic.node
+
+  components:
+    - id: G
+      model: basic.generator
+      parameters:
+        - id: cost
+          value: 100
+        - id: p_max
+          value: 100
+    - id: D
+      model: basic.demand
+      scenario-group: load
+      parameters:
+        - id: demand
+          time-dependent: true
+          scenario-dependent: true
+          value: loads
+
+  connections:
+    - component1: N
+      port1: injection_port
+      component2: D
+      port2: injection_port
+
+    - component1: N
+      port1: injection_port
+      component2: G
+      port2: injection_port
